requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
cryptography==44.0.3
exceptiongroup==1.3.0
fastapi==0.115.12
frozenlist==1.5.0
greenlet==3.1.1
h11==0.16.0
idna==3.10
jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.1.0
paramiko==3.5.1
propcache==0.2.0
pycparser==2.22
pydantic==2.8.2
pydantic-core==2.20.1
PyNaCl==1.5.0
python-dotenv==1.0.1
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.44.0
typing-extensions==4.13.2
uvicorn==0.33.0
yarl==1.15.2
